refactor(frontend): extract shared account form primitives
Move FormCard, FieldRow, ReadOnlyRow, ApiKeyPill, GeminiKeyField, and
analyzerModelLabel out of account.tsx into src/components/account-forms.tsx
so the upcoming Model Manager view (fs-23) can reuse them without
duplicating markup. No behavior change.

Refs #476
